requirements.txt
requests~=2.31.0
uvicorn~=0.24.0.post1
validators~=0.22.0
Pillow~=10.1.0
fastapi~=0.104.1
transformers~=4.35.2
pytest~=7.4.3